add file-based test and submission

+++ src/StarOJ.Data.Provider.SqlServer/TestCaseListProvider.cs
@@ -4,6 +4,7 @@ using System.Threading.Tasks;
 using StarOJ.Data.Provider.SqlServer.Models;
 using System.Linq;
 using StarOJ.Core.Problems;
+using System.IO;
 
 namespace StarOJ.Data.Provider.SqlServer
 {
@@ -12,27 +13,30 @@ namespace StarOJ.Data.Provider.SqlServer
         private readonly Workspace _workspace;
         private readonly OJContext _context;
         private readonly int _problemId;
+        private readonly bool _isSample;
 
-        public TestCaseListProvider(Workspace workspace, OJContext context, int problemId)
+        public TestCaseListProvider(Workspace workspace, OJContext context, int problemId, bool isSample)
         {
             _workspace = workspace;
             _context = context;
             _problemId = problemId;
+            _isSample = isSample;
         }
 
         public async Task Clear()
         {
-            var tests = (from x in _context.Tests where x.ProblemId == _problemId select x).ToArray();
+            var tests = (from x in _context.Tests where x.ProblemId == _problemId && x.IsSample == _isSample select x).ToArray();
             _context.Tests.RemoveRange(tests);
             await _context.SaveChangesAsync();
         }
 
         public async Task<ITestCaseProvider> Create(TestCaseMetadata metadata)
         {
-            TestCase empty = new TestCase { ProblemId = _problemId };
+            TestCase empty = new TestCase { ProblemId = _problemId, IsSample = _isSample };
             _context.Tests.Add(empty);
             await _context.SaveChangesAsync();
             var res = new TestCaseProvider(_workspace, _context, empty);
+            Directory.CreateDirectory(res.GetRoot());
             await res.SetMetadata(metadata);
             return res;
         }
@@ -46,8 +50,10 @@ namespace StarOJ.Data.Provider.SqlServer
         {
             int _id = int.Parse(id);
             var item = await _context.Tests.FindAsync(_id);
-            if (item != null && item.ProblemId == _problemId)
+            if (item != null && item.ProblemId == _problemId && item.IsSample == _isSample)
             {
+                var prov = new TestCaseProvider(_workspace, _context, item);
+                Directory.Delete(prov.GetRoot(), true);
                 _context.Tests.Remove(item);
                 await _context.SaveChangesAsync();
             }
@@ -57,24 +63,22 @@ namespace StarOJ.Data.Provider.SqlServer
         {
             int _id = int.Parse(id);
             var item = await _context.Tests.FindAsync(_id);
-            if (item == null || item.ProblemId != _problemId)
+            if (item != null && item.ProblemId == _problemId && item.IsSample == _isSample)
             {
-                return null;
+                return new TestCaseProvider(_workspace, _context, item); 
             }
             else
             {
-                return new TestCaseProvider(_workspace, _context, item);
+                return null;
             }
         }
 
         public Task<IEnumerable<ITestCaseProvider>> GetAll()
         {
-            var test = (from x in _context.Tests where x.ProblemId == _problemId select x).ToList();
+            var test = (from x in _context.Tests where x.ProblemId == _problemId && x.IsSample == _isSample select x).ToList();
             List<ITestCaseProvider> res = new List<ITestCaseProvider>();
             foreach (var v in test)
-            {
                 res.Add(new TestCaseProvider(_workspace, _context, v));
-            }
             return Task.FromResult((IEnumerable<ITestCaseProvider>)res);
         }
 
@@ -82,7 +86,7 @@ namespace StarOJ.Data.Provider.SqlServer
         {
             int _id = int.Parse(id);
             var item = await _context.Tests.FindAsync(_id);
-            if (item != null && item.ProblemId == _problemId)
+            if (item != null && item.ProblemId == _problemId && item.IsSample == _isSample)
             {
                 return true;
             }
